# apps/ai-broker-web/lib/openapi/investment-agent-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "AI Broker Investing Agent API",
    "version": "1.0.0",
    "description": "AI-powered multi-agent trading system API for comprehensive stock market analysis and automated trading decisions",
    "contact": {
      "name": "API Support",
      "url": "https://github.com/vtempest/investment-prediction-agent"
    }
  },
  "servers": [
    {
      "url": "https://autoinvestment.broker/api",
      "description": "Production server"
    },
    {
      "url": "http://localhost:3000/api",
      "description": "Development server"
    }
  ],
...
  "paths": {
...
    "/stocks/quote/{symbol}": {
      "get": {
        "tags": [
          "Stocks"
        ],
        "summary": "Get stock quote",
        "description": "Get current price and details for a stock symbol",
        "parameters": [
          {
            "name": "symbol",
            "in": "path",
            "required": true,
            "description": "Stock symbol (e.g., AAPL)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StockQuote"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "StockQuote": {
        "type": "object",
        "properties": {
          "symbol": {
            "type": "string"
          },
          "price": {
            "type": "number"
          },
          "open": {
            "type": "number"
          },
          "high": {
            "type": "number"
          },
          "low": {
            "type": "number"
          },
          "volume": {
            "type": "number"
          },
          "marketCap": {
            "type": "number"
          },
          "pe": {
            "type": "number"
          }
        }
      },
...
    }
  }
}